refactor: correctPath
