+++ src/commands/brotli/compress.js
@@ -1,13 +1,15 @@
-import path from 'path';
-import store from '../../store/store.js';
+import { open } from 'fs/promises';
 import { BROTLI_ACTIONS } from '../../constants/brotliActions.js';
 import { brotli } from '../../helpers/index.js';
+import { correctPath } from '../../helpers/correctPath.js';
 
 export const compress = async (filePath, destinationPath) => {
-  const readFile = path.resolve(store.currentPath, filePath);
-  const writeFile = path.resolve(store.currentPath, destinationPath);
+  const readFile = correctPath(filePath);
+  const writeFile = correctPath(destinationPath);
 
   try {
+    const file = await open(readFile, 'r');
+    file.close();
     await brotli(readFile, writeFile, BROTLI_ACTIONS.compress);
   } catch (error) {
     console.log(error.message);
