--- src/helpers/index.js
@@ -1,5 +1,0 @@
-import { copyFile } from './copyFile.js';
-import { parseInput } from './parseInput.js';
-import { brotli } from './brotli.js';
-
-export { copyFile, parseInput, brotli };
